feat: switched to ull int

+++ include/simple-logging.hh
@@ -5,9 +5,9 @@
 #include <ratio>
 #include <string>
 
-long int log_milli(std::string input)
+unsigned long long int log_milli(std::string input)
 {
-    long int time = std::chrono::duration_cast< std::chrono::milliseconds >(
+    unsigned long long int time = std::chrono::duration_cast< std::chrono::milliseconds >(
         std::chrono::system_clock::now().time_since_epoch()
     ).count();
 
@@ -16,9 +16,9 @@ long int log_milli(std::string input)
     return time;
 }
 
-long int log_milli(char* input)
+unsigned long long int log_milli(char* input)
 {
-    long int time = std::chrono::duration_cast< std::chrono::milliseconds >(
+    unsigned long long int time = std::chrono::duration_cast< std::chrono::milliseconds >(
         std::chrono::system_clock::now().time_since_epoch()
     ).count();
 
@@ -27,9 +27,9 @@ long int log_milli(char* input)
     return time;
 }
 
-long int log_micro(std::string input)
+unsigned long long int log_micro(std::string input)
 {
-    long int time = std::chrono::duration_cast< std::chrono::microseconds >(
+    unsigned long long int time = std::chrono::duration_cast< std::chrono::microseconds >(
         std::chrono::system_clock::now().time_since_epoch()
     ).count();
 
@@ -38,9 +38,9 @@ long int log_micro(std::string input)
     return time;
 }
 
-long int log_micro(char* input)
+unsigned long long int log_micro(char* input)
 {
-    long int time = std::chrono::duration_cast< std::chrono::microseconds >(
+    unsigned long long int time = std::chrono::duration_cast< std::chrono::microseconds >(
         std::chrono::system_clock::now().time_since_epoch()
     ).count();
 
@@ -49,9 +49,9 @@ long int log_micro(char* input)
     return time;
 }
 
-long int log_nano(std::string input)
+unsigned long long int log_nano(std::string input)
 {
-    long int time = std::chrono::duration_cast< std::chrono::nanoseconds >(
+    unsigned long long int time = std::chrono::duration_cast< std::chrono::nanoseconds >(
         std::chrono::system_clock::now().time_since_epoch()
     ).count();
 
@@ -60,9 +60,9 @@ long int log_nano(std::string input)
     return time;
 }
 
-long int log_nano(char* input)
+unsigned long long int log_nano(char* input)
 {
-    long int time = std::chrono::duration_cast< std::chrono::nanoseconds >(
+    unsigned long long int time = std::chrono::duration_cast< std::chrono::nanoseconds >(
         std::chrono::system_clock::now().time_since_epoch()
     ).count();
